Lock when writing to keymapper cache and memory concerns
Blocks should be copied to heap
http://stackoverflow.com/questions/14683746/why-should-we-copy-blocks-rather-than-retain

Sometimes EXC_BAD_ACCESS occurs when multiple threads are accessing the same JSONKeyMapper. The spin lock solves this.

Keeping a strongSelf through the block also avoids the possibility of self being nil between check and return
```
            if (strongSelf.toModelMap[keyName]) {
                return strongSelf.toModelMap[keyName];
            }
```
(using weakself there's a possibility that self is not nil in the if-test and is nil in the return statement)