Add module map file to targets other than iOS (#583)
In #482 a module map file was added that fixed an issue when installing
JSONModel with Carthage. But it seems to me that the fix was only applied
to the iOS target (or maybe the other targets were only added later?),
which means I can still see the issue when trying to install on macOS:

    <module-includes>:1:1: error: umbrella header for module 'JSONModel' does not include header 'JSONAPI.h' [-Werror,-Wincomplete-umbrella]
    #import "Headers/JSONModel.h"
    ^
    <module-includes>:1:1: error: umbrella header for module 'JSONModel' does not include header 'JSONHTTPClient.h' [-Werror,-Wincomplete-umbre
    <module-includes>:1:1: error: umbrella header for module 'JSONModel' does not include header 'JSONModel+networking.h' [-Werror,-Wincomplete
    <module-includes>:1:1: error: umbrella header for module 'JSONModel' does not include header 'JSONModelLib.h' [-Werror,-Wincomplete-umbrell
    4 errors generated.

This change adds the module map from #482 to all targets.